+++ MaterialDesignExtensions/Controllers/FileSystemController.cs
@@ -40,6 +40,7 @@ namespace MaterialDesignExtensions.Controllers
         private bool m_showSystemFilesAndDirectories;
         private IList<IFileFilter> m_fileFilters;
         private IFileFilter m_fileFilterToApply;
+        private bool m_forceFileExtensionOfFileFilter;
 
         /// <summary>
         /// The current directory shown in the control.
@@ -286,6 +287,27 @@ namespace MaterialDesignExtensions.Controllers
         }
 
         /// <summary>
+        /// Forces the possible file extension of the selected file filter for new filenames.
+        /// </summary>
+        public bool ForceFileExtensionOfFileFilter
+        {
+            get
+            {
+                return m_forceFileExtensionOfFileFilter;
+            }
+
+            set
+            {
+                if (m_forceFileExtensionOfFileFilter != value)
+                {
+                    m_forceFileExtensionOfFileFilter = value;
+
+                    OnPropertyChanged(nameof(ForceFileExtensionOfFileFilter));
+                }
+            }
+        }
+
+        /// <summary>
         /// The special directories (e.g. music directory) of the user.
         /// </summary>
         public List<SpecialDirectory> SpecialDirectories
@@ -367,6 +389,7 @@ namespace MaterialDesignExtensions.Controllers
             m_showSystemFilesAndDirectories = false;
             m_fileFilters = null;
             m_fileFilterToApply = null;
+            m_forceFileExtensionOfFileFilter = false;
         }
 
         /// <summary>
@@ -582,6 +605,54 @@ namespace MaterialDesignExtensions.Controllers
 
             // important: create a new DirectoryInfo instance, because the Exists property will not be updated by the Create() method
             SelectDirectory(new DirectoryInfo(fullDirectoryName));
+        }
+
+        /// <summary>
+        /// Builds a full filename for the specified filename inside the current directory.
+        /// </summary>
+        /// <param name="newFilename">The filename to append to the current directory</param>
+        /// <returns></returns>
+        public string BuildFullFileNameForInCurrentDirectory(string newFilename)
+        {
+            string filename = null;
+
+            if (!string.IsNullOrWhiteSpace(newFilename) && m_currentDirectory != null)
+            {
+                string directory = m_currentDirectory.FullName;
+
+                if (directory != null && !directory.EndsWith(@"\") && !directory.EndsWith("/"))
+                {
+                    directory = $@"{directory}\";
+                }
+
+                filename = directory + newFilename.Trim();
+
+                // ensure that the full filename has a file extension out of the selected file filter
+                if (m_forceFileExtensionOfFileFilter && m_fileFilterToApply != null && !m_fileFilterToApply.IsMatch(filename))
+                {
+                    IEnumerable<string> fileExtensions = FileFilterHelper.GetFileExtensionsFromFilter(m_fileFilterToApply);
+
+                    if (fileExtensions != null && fileExtensions.Any())
+                    {
+                        fileExtensions = fileExtensions.Select(fileExtension => fileExtension.ToLower());
+                        string lowerCaseFilename = filename.ToLower();
+
+                        bool hasWrongFileExtension = !fileExtensions.Any(fileExtension => lowerCaseFilename.EndsWith($".{fileExtension}"));
+
+                        if (hasWrongFileExtension)
+                        {
+                            if (!filename.EndsWith("."))
+                            {
+                                filename = $"{filename}.";
+                            }
+
+                            filename = filename + fileExtensions.First();
+                        }
+                    }
+                }
+            }
+
+            return filename;
         }
 
         private bool AreObjectsEqual(object o1, object o2)